feat: move to fully client side rendering
