--- apps/cms/src/app/(main)/[workspace]/(workspace)/settings/page.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "next/navigation";
-
-async function Page() {
-  return redirect("/settings/general");
-}
-
-export default Page;
